+++ docs/changelog.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 Each revision is versioned by the date of the revision.
 
+### 2025-09-23
+
+- Support updated auth-proxy relation.
+
 ### 2025-09-18
 
 - Fixed issue where naming the application deployment to "jenkins" would not add relation data with